Create README.md
This repository contains all my solutions to the exercises in the Data Analysis with Python course on freeCodeCamp. 


In the Data Analysis with Python Certification, you'll learn the fundamentals of data analysis with Python. 

You'll know how to read data from sources like CSVs and SQL, and 
how to use libraries like Numpy, Pandas, Matplotlib, and Seaborn  and other libraries to process and visualize data.