Create README.md
This repository contains all my solutions to the exercises in the College Algebra with Python course on freeCodeCamp.

This course is designed as a one-semester college course. 

It consists of instructional videos, Google Colaboratory notebooks to follow along interactively, assignments, and challenging projects.

You will gain a full understanding of Algebra and how to write Python code to solve Algebra problems.

You will also build your own Algebra Colab notebook that you will be able to use as your custom calculator. 

This course (and the code you write here) will give you the foundation for a deeper math and data science understanding.